requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
coveralls==2.0.0
Django==3.0.5
django-cors-headers==3.2.1
django-render-block==0.6
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docopt==0.6.2
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==4.0.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.3
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
random-password-generator==2.1.0
redis==3.4.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
